master2/METHODOLOGIE_TESTS/5_TD/methodologie-du-test/requirements.txt
ruff
pytest
python-dotenv
pandas
ulid-py
